requirements.txt
pytest==8.3.4
psycopg2-binary==2.9.10
pymysql==1.1.1
redis==5.2.1
hiredis==3.1.0
python-dotenv==1.0.1